generateFolderAndIndex split + fixed bad defaultStructureItem
