+++ .gitignore
@@ -8,5 +8,5 @@ TESTS_SRC
 STRUCTURE_TESTS_SRC
 logs.txt
 /plugins
-.DEV_FOLDER
+.DEV*
 architect.config.json
